整合canal，实现es和mysql全量、增量数据同步

+++ elasticsearch-demo/src/main/java/com/free/es/utils/EsUtil.java
@@ -2,58 +2,76 @@ package com.free.es.utils;
 
 import cn.hutool.core.bean.BeanUtil;
 import cn.hutool.json.JSONUtil;
-import lombok.AllArgsConstructor;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.StringUtils;
 import org.elasticsearch.action.admin.indices.delete.DeleteIndexRequest;
 import org.elasticsearch.action.delete.DeleteRequest;
+import org.elasticsearch.action.get.GetRequest;
+import org.elasticsearch.action.get.GetResponse;
 import org.elasticsearch.action.index.IndexRequest;
+import org.elasticsearch.action.index.IndexResponse;
 import org.elasticsearch.action.search.SearchRequest;
 import org.elasticsearch.action.search.SearchResponse;
 import org.elasticsearch.action.support.master.AcknowledgedResponse;
 import org.elasticsearch.action.update.UpdateRequest;
-import org.elasticsearch.client.HttpAsyncResponseConsumerFactory;
 import org.elasticsearch.client.RequestOptions;
 import org.elasticsearch.client.RestHighLevelClient;
 import org.elasticsearch.client.indices.CreateIndexRequest;
 import org.elasticsearch.client.indices.CreateIndexResponse;
-import org.elasticsearch.common.settings.Settings;
+import org.elasticsearch.client.indices.GetIndexRequest;
+import org.elasticsearch.common.Strings;
 import org.elasticsearch.common.xcontent.XContentType;
 import org.elasticsearch.index.query.*;
 import org.elasticsearch.search.builder.SearchSourceBuilder;
+import org.elasticsearch.search.fetch.subphase.FetchSourceContext;
 import org.elasticsearch.search.sort.SortOrder;
 import org.springframework.stereotype.Component;
 
 import java.io.IOException;
+import java.util.Map;
 
 @Slf4j
 @Component
-@AllArgsConstructor
-public class EsUtil<T> {
+@RequiredArgsConstructor
+public class EsUtil {
 
-    private RestHighLevelClient client;
-
-    private static final RequestOptions COMMON_OPTIONS;
-
-    static {
-        RequestOptions.Builder builder = RequestOptions.DEFAULT.toBuilder();
-        // 默认缓冲限制为100MB，此处修改为30MB。
-        builder.setHttpAsyncResponseConsumerFactory(new HttpAsyncResponseConsumerFactory.HeapBufferedResponseConsumerFactory(30 * 1024 * 1024));
-        COMMON_OPTIONS = builder.build();
-    }
+    private final RestHighLevelClient client;
 
     /**
      * 创建索引库
      *
      * @param index
      */
-    public void createIndexRequest(String index) {
-        CreateIndexRequest createIndexRequest = new CreateIndexRequest(index).settings(Settings.builder().put("index.number_of_shards", 3).put("index.number_of_replicas", 0));
+    public boolean createIndexRequest(String index) {
+        if (isIndexExist(index)) {
+            log.error("Index is exits!");
+            return false;
+        }
+        CreateIndexRequest request = new CreateIndexRequest(index);
         try {
-            CreateIndexResponse response = client.indices().create(createIndexRequest, COMMON_OPTIONS);
-            log.info(" 所有节点确认响应 : {}", response.isAcknowledged());
-            log.info(" 所有分片的复制未超时 :{}", response.isShardsAcknowledged());
+            CreateIndexResponse response = client.indices().create(request, RequestOptions.DEFAULT);
+            log.info("创建索引{}成功", index);
+            return response.isAcknowledged();
         } catch (IOException e) {
-            log.error("创建索引库【{}】失败", index, e);
+            log.error("创建索引库 {" + index + "}失败", e);
+            return false;
+        }
+    }
+
+    /**
+     * 判断索引是否存在
+     *
+     * @param index
+     * @return
+     */
+    public boolean isIndexExist(String index) {
+        GetIndexRequest request = new GetIndexRequest(index);
+        try {
+            return client.indices().exists(request, RequestOptions.DEFAULT);
+        } catch (IOException e) {
+            log.error("索引查询失败", e);
+            return false;
         }
     }
 
@@ -62,28 +80,37 @@ public class EsUtil<T> {
      *
      * @param index
      */
-    public void deleteIndexRequest(String index) {
+    public boolean deleteIndexRequest(String index) {
+        if (!isIndexExist(index)) {
+            log.error("Index is not exits!");
+            return false;
+        }
+        //删除索引请求
         DeleteIndexRequest request = new DeleteIndexRequest(index);
         try {
-            AcknowledgedResponse response = client.indices().delete(request, COMMON_OPTIONS);
-            System.out.println(response.isAcknowledged());
+            //执行客户端请求
+            AcknowledgedResponse delete = client.indices().delete(request, RequestOptions.DEFAULT);
+            log.info("删除索引库{}成功", index);
+            return delete.isAcknowledged();
         } catch (IOException e) {
-            log.error("删除索引库【{}】失败", index, e);
+            log.error("删除索引库 {" + index + "}失败", e);
+            return false;
         }
     }
 
     /**
-     * 更新索引文档
+     * 通过id更新索引文档
      *
      * @param index
      * @param id
      * @param object
      */
-    public void updateRequest(String index, String id, Object object) {
+    public void updateRequestById(String index, String id, Object object) {
         UpdateRequest updateRequest = new UpdateRequest(index, id);
         updateRequest.doc(BeanUtil.beanToMap(object), XContentType.JSON);
         try {
-            client.update(updateRequest, COMMON_OPTIONS);
+            client.update(updateRequest, RequestOptions.DEFAULT);
+            log.info("索引为: {}, id为: {}, 更新数据成功", index, id);
         } catch (IOException e) {
             log.error("更新索引文档 {" + index + "} 数据 {" + object + "} 失败", e);
         }
@@ -99,24 +126,74 @@ public class EsUtil<T> {
     public void insertRequest(String index, String id, Object object) {
         IndexRequest indexRequest = new IndexRequest(index).id(id).source(BeanUtil.beanToMap(object), XContentType.JSON);
         try {
-            client.index(indexRequest, COMMON_OPTIONS);
+            IndexResponse response = client.index(indexRequest, RequestOptions.DEFAULT);
+            log.info("添加数据成功 索引为: {}, response 状态: {}, id为: {}", index, response.status().getStatus(), response.getId());
         } catch (IOException e) {
             log.error("创建索引文档 {" + index + "} 数据 {" + object + "} 失败", e);
         }
     }
 
     /**
-     * 删除索引文档
+     * 根据id删除索引文档
      *
      * @param index
      * @param id
      */
-    public void deleteRequest(String index, String id) {
+    public void deleteRequestById(String index, String id) {
         DeleteRequest deleteRequest = new DeleteRequest(index, id);
         try {
-            client.delete(deleteRequest, COMMON_OPTIONS);
+            client.delete(deleteRequest, RequestOptions.DEFAULT);
+            log.info("索引为: {}, id为: {}删除数据成功", index, id);
         } catch (IOException e) {
             log.error("删除索引文档 {" + index + "} 数据id {" + id + "} 失败", e);
+        }
+    }
+
+
+    /**
+     * 通过ID获取数据
+     *
+     * @param index  索引库
+     * @param id     数据ID
+     * @param fields 需要显示的字段，逗号分隔（缺省为全部字段）
+     * @return
+     */
+    public Map<String, Object> searchDataById(String index, String id, String fields) {
+        GetRequest request = new GetRequest(index, id);
+        if (StringUtils.isNotEmpty(fields)) {
+            //只查询特定字段。如果需要查询所有字段则不设置该项。
+            request.fetchSourceContext(new FetchSourceContext(true, fields.split(","), Strings.EMPTY_ARRAY));
+        }
+        GetResponse response = null;
+        try {
+            response = client.get(request, RequestOptions.DEFAULT);
+            Map<String, Object> map = response.getSource();
+            //为返回的数据添加id
+            map.put("id", response.getId());
+            return map;
+        } catch (IOException e) {
+            log.error("获取数据失败", e);
+            return null;
+        }
+    }
+
+    /**
+     * 通过ID判断文档是否存在
+     *
+     * @param index 索引，类似数据库
+     * @param id    数据ID
+     * @return
+     */
+    public boolean existsById(String index, String id) {
+        GetRequest request = new GetRequest(index, id);
+        //不获取返回的_source的上下文
+        request.fetchSourceContext(new FetchSourceContext(false));
+        request.storedFields("_none_");
+        try {
+            return client.exists(request, RequestOptions.DEFAULT);
+        } catch (IOException e) {
+            log.error("判断失败", e);
+            return false;
         }
     }
 
@@ -125,23 +202,13 @@ public class EsUtil<T> {
      *
      * @param index
      * @param queryBuilder
-     * @param sort
      */
     public SearchResponse search(String index, QueryBuilder queryBuilder) {
         SearchRequest searchRequest = new SearchRequest(index);
         SearchSourceBuilder searchSourceBuilder = new SearchSourceBuilder();
         searchSourceBuilder.query(queryBuilder);
-        //bool符合查询
-        //BoolQueryBuilder boolQueryBuilder = new BoolQueryBuilder()
-        //        .filter(QueryBuilders.matchQuery("name", "盖伦"))
-        //        .must(QueryBuilders.matchQuery("desc", "部落"))
-        //        .should(QueryBuilders.matchQuery("realName", "光辉"));
-
-        //分页
-        //searchSourceBuilder.from(1).size(2);
         // 排序
         searchSourceBuilder.sort("createTime", SortOrder.DESC);
-
         ////误拼写时的fuzzy模糊搜索方法 2表示允许的误差字符数
         //QueryBuilders.fuzzyQuery("title", "ceshi").fuzziness(Fuzziness.build("2"));
         searchRequest.source(searchSourceBuilder);
@@ -149,9 +216,10 @@ public class EsUtil<T> {
         System.out.println(JSONUtil.parseObj(searchSourceBuilder.toString()).toStringPretty());
         SearchResponse searchResponse = null;
         try {
-            searchResponse = client.search(searchRequest, COMMON_OPTIONS);
+            searchResponse = client.search(searchRequest, RequestOptions.DEFAULT);
         } catch (IOException e) {
             e.printStackTrace();
+            log.error("搜索数据失败", e);
         }
         return searchResponse;
     }
